requirements.txt
click==8.0.1
google-api-core==2.0.1
google-api-python-client==2.23.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
haxor==1.2.3
httplib2==0.19.1
oauthlib==3.1.1
pytest==6.2.4
python-decouple==3.5
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==1.3.0
